# skill/src/main/java/com/capgemini/talent/skill/api/entity/TalentPoolCandidate.java
package com.capgemini.talent.skill.api.entity;

import jakarta.persistence.Entity;
import jakarta.persistence.GeneratedValue;
import jakarta.persistence.GenerationType;
import jakarta.persistence.Id;
import lombok.Data;

@Entity
@Data
public class TalentPoolCandidate {

	@Id
	@GeneratedValue(strategy = GenerationType.IDENTITY)
	private Long id;

	private String name;

	private String grade;

	private String technologies;

	private String role;

	private String candidateCertification;

	private String languages;

}