# skill/src/main/resources/APISpecification.yml
openapi: 3.0.3
info:
  title: Talent Pool API
  description: API to manage talent pool candidates.
  version: 1.0.0
  contact:
    name: Rupesh S.A
    email: rupesh.sukale@capgemini.com

servers:
  - url: https://api.talentpool.com/v1
    description: Production Server

paths:
  /candidates:
    get:
      summary: Get all candidates
      description: Fetch a list of all talent pool candidates.
      operationId: getCandidates
      security:
        - OAuth2: [read]
      responses:
        '200':
          description: A list of talent pool candidates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TalentPoolCandidate'
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error

    post:
      summary: Create a new candidate
      description: Add a new talent pool candidate to the database.
      operationId: createCandidate
      security:
        - OAuth2: [write]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TalentPoolCandidate'
      responses:
        '201':
          description: Candidate created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TalentPoolCandidate'
        '400':
          description: Bad Request (Invalid data)
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error

  /candidates/{id}:
    get:
      summary: Get a candidate by ID
      description: Fetch a specific talent pool candidate by ID.
      operationId: getCandidateById
      security:
        - OAuth2: [read]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Candidate found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TalentPoolCandidate'
        '404':
          description: Candidate not found
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error

    put:
      summary: Update a candidate
      description: Update an existing talent pool candidate.
      operationId: updateCandidate
      security:
        - OAuth2: [write]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TalentPoolCandidate'
      responses:
        '200':
          description: Candidate updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TalentPoolCandidate'
        '400':
          description: Bad Request (Invalid data)
        '404':
          description: Candidate not found
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error

    delete:
      summary: Delete a candidate
      description: Delete a talent pool candidate by ID.
      operationId: deleteCandidate
      security:
        - OAuth2: [write]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Candidate deleted successfully
        '404':
          description: Candidate not found
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error

components:
  securitySchemes:
    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://talentpool/authorize
          tokenUrl: https://talentpool/token
          scopes:
            read: Grants read access
            write: Grants write access

  schemas:
    TalentPoolCandidate:
      type: object
      properties:
        id:
          type: long
          format: int64
        name:
          type: string
        certifications:
          type: string
        grade:
          type: string
        languages:
          type: string
        role:
          type: string
        technologies:
          type: string
      required:
        - name
        - role
        - technologies


  responses:
    EmployeeNotFound:
      description: Employee not found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Employee not found"